Reactive layout detail page
